Moved traits to a place that makes a little more sense.
